Add avatar compositing pipeline to nitro-swift
Implements the figure-string-to-composited-layers pipeline for a static
(non-animated) avatar pose:

- FigureSetData/Palette/SetType/FigurePartSet/FigurePart: the figuredata.json
  clothing catalog.
- PartSetsData/ActivePartSet/PartDefinition: which figure parts an action
  activates, and the mirrored left/right hand-item slots.
- AvatarModelGeometry: config-driven body-part depth sort, reusing the
  Y-rotation/camera-distance algorithm added for this port, including a
  faithfully-replicated quirk in the original (AvatarModelGeometry.getParts
  rotates by the raw 0-7 direction index rather than the resolved compass
  angle that AvatarModelGeometry.getBodyPartsAtAngle uses).
- AvatarStructure: figure+action -> visible AvatarImagePartContainer layers.
- AssetAliasCollection: alias-chain asset name resolution for avatar parts.
- AvatarCompositor: asset naming with its 4-step fallback chain, the
  gesture/held-item mirror-in-place vs. direction-substitution logic, and the
  union-image placement math (worked out by hand from the Pixi Matrix/regPoint
  arithmetic in AvatarImageCache) that positions every layer within the
  avatar canvas.
- AvatarDefaults: loads the client's built-in default geometry/part-set/action
  tables, extracted verbatim from their TS sources into bundled JSON resources
  via a small esbuild-based script (not hand-transcribed, to avoid transcription
  errors in ~1800 lines of coordinate data) - includes a real base figure
  (body+head+swimsuit), so a placeholder avatar renders correctly even before
  real figuredata.json is injected.
- NitroRendererKit.AvatarNode: SpriteKit adapter; RoomScene.placeAvatar wires
  it into the room scene alongside furniture.

Out of scope for this pass (documented in the README): keyframe animation
playback, the action precedence/combination system beyond a single action,
and gesture/FX-injected layers - only the "Stand" posture is wired up.

Also fixes several Double/CGFloat argument-type mismatches in the furniture
and asset code found while writing this (CGPoint/CGRect initializers need
explicit CGFloat(...) conversions in Swift; plain Double/Int values don't
convert implicitly).

+++ packages/nitro-swift/Sources/NitroAssets/GraphicAssetPalette.swift
@@ -54,7 +54,7 @@ public final class GraphicAssetPalette {
                 space: colorSpace, bitmapInfo: CGImageAlphaInfo.premultipliedLast.rawValue
             ) else { return nil }
 
-            ctx.draw(cgImage, in: CGRect(x: 0, y: 0, width: width, height: height))
+            ctx.draw(cgImage, in: CGRect(x: 0, y: 0, width: CGFloat(width), height: CGFloat(height)))
 
             let pixels = rawBuffer.bindMemory(to: UInt8.self)
             var index = 0
